Corrigidos includes

+++ projeto_1/src/i2c_bme.c
@@ -1,15 +1,4 @@
-#include <string.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <unistd.h>
-#include <sys/types.h>
-#include <fcntl.h>
-
-#include "bme280.h"
 #include "i2c_bme.h"
-
-#include <linux/i2c-dev.h>
-#include <sys/ioctl.h>
 
 void _user_delay_us(uint32_t period, void *intf_ptr);
 int8_t _user_i2c_read(uint8_t reg_addr, uint8_t *data, uint32_t len, void *intf_ptr);
@@ -99,7 +88,7 @@ int8_t read_temperature_i2c(struct bme280_dev *dev, float *temp)
         return rslt;
     }
 
-    *temp = comp_data.temperature;
+    *temp = (float)comp_data.temperature;
 
     print_sensor_data(&comp_data);
